% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateReformatter.R
\name{dateReformatter}
\alias{dateReformatter}
\title{dateReformatter function}
\usage{
dateReformatter(datV)
}
\arguments{
\item{datV}{character vector of dates.}
}
\value{
Returns the vector of standardized and corrected dates.
}
\description{
This function helps to standartize and correct misspellings of date entries.
}
\examples{
library(HooverArchives)

datesV<- c("1914:Aug 2 - 20; 1918:NOv 18",
           "1941:August-September 5, 1943 :Aug-1944:July")

reformated.data <- dateReformatter(datesV)
reformated.data
}
