% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromLATtoCYR.R
\name{fromLATtoCYR}
\alias{fromLATtoCYR}
\title{Latin to Cyrillic conversion function}
\usage{
fromLATtoCYR(
  mdat,
  LARU = TRUE,
  RURU = FALSE,
  EnglishDetection = TRUE,
  EnglishLength = NULL,
  RussianCorrection = FALSE,
  SensitivityThreshold = 0.1
)
}
\arguments{
\item{mdat}{character vector to be back-transliterated to Cyrillic.}

\item{LARU}{rules of tranliteration from transliterated Cyrillic to original Cyrillic (the rules are listed in the file "transliterationLARU.csv").}

\item{RURU}{rules to correct transliterated original Cyrillic (the rules are listed in the file "transliterationRURU.csv").}

\item{EnglishDetection}{if set to TRUE, the script avoids transliteration of words found in the English vocabulary (file: english.txt). If set to FALSE, only user defined stop words are used (file: stopwordsfile.csv).}

\item{EnglishLength}{threshold is set to ignore EnglishDectection words below given threshold.}

\item{RussianCorrection}{if set to TRUE, the script attempts to match every back-transliterated word with the Russian vocabulary (files: russian.txt and russian_surnames.txt).}

\item{SensitivityThreshold}{is used only if RussianCorrection==TRUE. It determines algorithm's sensitivity to mismatches (numbers closer to 0 define higher sensitivity to mismatches). SensitivityThreshold is set to 0.1 by default.}
}
\value{
Returns the vector of transliterated characters in Cyrillic.
}
\description{
This function helps to convert transliterated Cyrillic to original Cyrillic.
}
\examples{
library(HooverArchives)

dat<-c("Mezhdunarodnaia gazeta. Gl. redaktor: Iu. Zarechkin. Moscow, Russia. Semiweekly. 199?",
"DEN' UCHITELIA komissiia po obrazovaniiu ob''edineniia Iabloko",
"III-ii RIM vestnik Rossiiskogo patrioticheskogo dvizheniia. Redaktory: M. Artem'ev, V. Rugich. Moscow, Russia.")

converteddata <- fromLATtoCYR(dat, LARU=TRUE, RURU=FALSE, EnglishDetection=TRUE, EnglishLength=4)
}
