% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromSERIEStoFILES.R
\name{fromSERIEStoFILES}
\alias{fromSERIEStoFILES}
\title{fromSERIEStoFILES function}
\usage{
fromSERIEStoFILES(dat = NULL, issueDates = NULL, locale = "English")
}
\arguments{
\item{dat}{original dataframe.}

\item{issueDates}{Variables containing information.}

\item{locale}{system's locale (English by default).}
}
\value{
Returns the dataframe with the dataframe.
}
\description{
This function helps to obtain "Files" variable by splitting the issueDates into separate years.
}
\examples{
library(HooverArchives)

item_nodes<-list(path="(//c)|(//c01)|(//c02)|(//c03)|(//c04)|(//c05)|(//c06)|(//c07)|(//c08)",
                 nodes=c("primarynode", ".//unittitle", "./did//unitdate", "./did//unitdate",
                         "./did//language", "./did//abstract", ".//container",
                         "./did//container", "./scopecontent[@id]", ".//scopecontent/head",
                         ".//scopecontent/p", "./accessrestrict[@id]",
                         ".//accessrestrict/head", ".//accessrestrict/p", ".//note"),
                 types=c("attrs", "text", "text", "attrs", "attrs", "text", "text", "attrs", "attrs",
                         "text", "text", "attrs", "text", "text", "text"))

collection_nodes<-list(path="archdesc[@level='collection']",
                 nodes=c(".//unittitle", "./did//unitdate", "./did//language", "./did//abstract"),
                 types=c("text", "text", "attrs", "text"))

file_transf<-fromXMLtoCSV(system.file("rusnewspapers.xml", package="HooverArchives"), item_nodes, collection_nodes)

convdata<-fromSERIEStoFILES(file_transf, issueDates="note.text", locale="Russian")

#write.csv(convdata, "convdata_2012C30.csv")
}
