% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusnewspapers.R
\docType{data}
\name{rusnewspapers}
\alias{rusnewspapers}
\title{Soviet and post-Soviet independent publications collection}
\format{
XML
}
\description{
Soviet and post-Soviet independent publications collection, Hoover Institution Library & Archives, USA
}
\references{
See \href{https://oac.cdlib.org/findaid/ark:/13030/c8w66j67/entire_text/}{Inventory of the Soviet and post-Soviet independent publications collection}.
}
\keyword{Inventory}
\keyword{Soviet}
\keyword{and}
\keyword{collection}
\keyword{independent}
\keyword{of}
\keyword{post-Soviet}
\keyword{publications}
\keyword{the}
