% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCharSystem.R
\name{testCharSystem}
\alias{testCharSystem}
\title{testCharSystem function}
\usage{
testCharSystem(dat, addCharSys = NULL, markword = TRUE, autochange = FALSE)
}
\arguments{
\item{dat}{data vector}

\item{addCharSys}{the list of character blocks.  If not defined, c("Latin", "Cyrillic") are used.}

\item{markword}{if TRUE (default), detect the word containing anomalous  character and mark it. If FALSE, detect an anomalous character within a word and mark that character.}

\item{autochange}{if TRUE change characters based on the proposed coding rules in "charcodescheme.csv" (for more than one character "|" separator appears, for an unknown character that character is replaced with "?").}
}
\value{
Returns an altered data vector with anomalous words/characters/replacements surrounded by asterisks (*).
}
\description{
This function helps to detect characters originated from different Unicode blocks.
}
\examples{

library(HooverArchives)
library(stringi)

dat_vectorR <- c("\u0418\u043D\u0444\u043E\u0440\u043C\u0061\u0446\u0438\u044F", "\u0410\u0440\u0078\u0438\u0432\u044B")
dat_vector <- stri_unescape_unicode(dat_vectorR)

# Mark the word
testCharSystem(dat_vector, addCharSys=c("Latin", "Cyrillic"), autochange=FALSE, markword=TRUE)

# Mark anamolous character
testCharSystem(dat_vector, addCharSys=c("Latin", "Cyrillic"), autochange=FALSE, markword=FALSE)

# Replace anamolous character with correct character and mark it
testCharSystem(dat_vector, addCharSys=c("Latin", "Cyrillic"), autochange=TRUE, markword=FALSE)
}
