% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFILEStoSERIES.R
\name{fromFILEStoSERIES}
\alias{fromFILEStoSERIES}
\title{fromFILEStoSERIES function}
\usage{
fromFILEStoSERIES(
  dat = NULL,
  series_title = NULL,
  files = NULL,
  series_scope_note = NULL,
  series_date_range = NULL,
  scope_and_content = NULL,
  problems_notes = NULL,
  box_barcode = NULL,
  top_container = NULL,
  ...
)
}
\arguments{
\item{dat}{data frame}

\item{series_title}{variable name "Series title"}

\item{files}{variable name for "Files"}

\item{series_scope_note}{variable name for "Series scope note"}

\item{series_date_range}{variable name for "Hoover date range"}

\item{scope_and_content}{variable name for "Scope and content"}

\item{problems_notes}{variable name for "Problems/Notes"}

\item{box_barcode}{variable name for "Box Barcode"}

\item{top_container}{variable name for "Top container"}

\item{...}{other parameters}
}
\value{
Returns the vector of standardized and corrected dates.
}
\description{
This function helps to organize the Belgium data by adding extra "Series" row and reformatting data in a proper format.
}
\examples{
library(HooverArchives)
library(readxl)
library(xlsx)

#Load data and create indices
dat2.1<-read.xlsx(system.file("belgiumdata.xlsx", package="HooverArchives"), sheetIndex=1, header=FALSE, encoding = "utf-8")
dat2.1[]<-lapply(dat2.1, as.character)
colnames(dat2.1)<-as.character(dat2.1[3,])
dat2.1<-dat2.1[-(1:3),-c(1,14)];
dat2.1$indexW<-dat2.1$`Item title`

dat2.2<-read.xlsx(system.file("belgiumdata.xlsx", package="HooverArchives"), sheetIndex=2, header=TRUE, encoding = "utf-8")
dat2.2$indexW<- dat2.2$`Packet.Catalog.Title`

#Merge two dataframe using BuildIndex and Merge_data functions
index_matches<-buildIndex(dat2.1$indexW,dat2.2$indexW,
                          index_simplify=TRUE,
                          fuzzy_matching=TRUE,
                          index_hashing=FALSE)
mdat<-mergeData(dat2.1,dat2.2, index_matches)

#Use fromFILEStoSERIES() to add the Series row
coverted.dat<-fromFILEStoSERIES(dat=mdat,
                                series_title="Series title",
                                files="index",
                                series_scope_note="Series scope note",
                                series_date_range="Hoover date range",
                                scope_and_content="Scope.and.content",
                                problems_notes="Series scope note",
                                box_barcode="Box_Barcode",
                                top_container="Final.Box..")

datHarvard<-subset(coverted.dat, select=c("Group", "Title", "Hierarchical_Relationship",	"Processing_Information",
                                           "Description_Level",	"Date", "Top_Container_[indicator]", "Box_Barcode",
                                           "Scope_and_content"))
#Save file in xlsx to preserve diacritic characters
#write.xlsx(datHarvard, "convertedtoHarvardStyle.xlsx", sheetName = "HarvardStyle", col.names = TRUE)
}
